refactor: extract out SQLProblem component
