+++ assignments-portal/src/utils.js
@@ -1,0 +1,7 @@
+import markdownit from 'markdown-it'
+
+const md = markdownit();
+
+export function md2html(markdown) {
+    return md.render(markdown);
+}
